find virtual functions for CBaseEntity
virtual functions are found by replacing the VTable in the metamod with one that logs which function offset was called.

This only works for CBaseEntity for some reason. Probably the derived classes use multiple inheritance and so another vtable needs to be replaced somewhere else in the private data. Finding that will be challenging because the size of each base class isn't known, but it can probably still be brute forced with lots of time.